third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/unresolved-pseudoclass-expected.txt

CONSOLE ERROR: line 34: Uncaught TypeError: document.registerElement is not a function
Tests the :unresolved pseudoclass.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL document.querySelector("x-x:unresolved") should be [object HTMLElement]. Threw exception SyntaxError: Failed to execute 'querySelector' on 'Document': 'x-x:unresolved' is not a valid selector.
FAIL window.getComputedStyle(a).color should be rgb(0, 0, 222). Was rgb(0, 222, 0).
FAIL window.getComputedStyle(b).color should be rgb(0, 0, 222). Was rgb(0, 0, 0).
FAIL window.getComputedStyle(b).borderColor should be rgb(0, 0, 111). Was rgb(0, 0, 0).
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
